removed float fix since it did not fix

+++ database/migrations/2022_03_29_093023_create_dangerzones_table.php
@@ -17,8 +17,8 @@ class CreateDangerzonesTable extends Migration
             $table->id();
             $table->integer('fishpond_id');
             $table->string('data_type');
-            $table->float('min');
-            $table->float('max');
+            $table->integer('min');
+            $table->integer('max');
             $table->timestamps();
         });
     }
